Remove youtube "closed caption exists already" post on other subs
This is a nice-to-have, and it seems Reddit thinks it's pretty spammy if
we do that. Maybe. Either way, it has consistently resulted in a 403
response from reddit for videos where u/tor would post such a message.
Consistent means it has happened literally for the past month (Dec 18 -
Jan 17)